refactor: Seperate UI and domain logic
